backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.4.0
xgboost==2.0.3
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0